requirements-dev.txt
-r requirements.txt
black
pylint
pylint-django
mypy
pytest
pytest-xdist
coverage